feat: add hoc to display side menu
